lamcal done

+++ src/name2.f90
@@ -61,7 +61,7 @@ module fsqu
       use stel_kinds, only: dp
       use name1,      only: nsd
       implicit none
-      real(kind=dp) :: mns
+      integer       :: mns
       real(kind=dp) :: fnorm
       real(kind=dp) :: fsqr
       real(kind=dp) :: fsqz
